Optimize simple weekly case
Using a different strategy for the simple weekly case with no
additional filters results in a 40-50% faster expansion for those rules,
when expanded over a short timespan of a month (I expect that the
speedup would be even larger when expanding over long periods of time,
as in those cases the time taken to expand is proportionally larger than
the initialization time of parsing options, etc).

I think it makes sense to start with this case for now and gain some
confidence that it doesn't miss any edge cases, but this strategy
(identifying simple cases and handing them separately than the cases
that require filters) might be a good one to apply here in general.

Not ready to merge this yet, as I want to update the specs for the frequencies
if this approach looks good, but wanted to get a PR out to collect feedback.