Fix bug with simple weekly frequency when ENV['TZ'] does not match date_start timezone
When comparing days of the week, it's important that both the
`@current_date` and the `date_start` are in the same timezone. However,
this wasn't the case, leading to the first occurrence not being returned
if the _day of week_ was different between the ENV['TZ'] and the
date_start timezone.